+++ rainier-core/src/main/scala/com/stripe/rainier/sampler/HMC.scala
@@ -8,15 +8,16 @@ final case class HMC(nSteps: Int) extends Sampler {
              warmupIterations: Int,
              iterations: Int,
              keepEvery: Int)(implicit rng: RNG): List[Array[Double]] = {
-    val chain = HamiltonianChain(density.variables, density)
+    val lf = LeapFrog(density.variables.toList, density)
+    val params = lf.initialize
     val stepSize =
-      DualAvg.findStepSize(chain, 0.65, nSteps, warmupIterations)
+      DualAvg.findStepSize(lf, params, 0.65, nSteps, warmupIterations)
     val buf = new ListBuffer[Array[Double]]
     var i = 0
     while (i < iterations) {
-      chain.step(stepSize, nSteps)
+      lf.step(params, nSteps, stepSize)
       if (i % keepEvery == 0)
-        buf += chain.variables
+        buf += lf.variables(params)
       i += 1
     }
     buf.toList
